+++ library/azure_rm_sqlserver.py
@@ -30,9 +30,6 @@ options:
         description:
             - The name of the server.
         required: True
-    tags:
-        description:
-            - Resource tags.
     location:
         description:
             - Resource location.
@@ -125,10 +122,6 @@ class AzureRMServers(AzureRMModuleBase):
                 type='str',
                 required=True
             ),
-            tags=dict(
-                type='dict',
-                required=False
-            ),
             location=dict(
                 type='str',
                 required=False
@@ -177,18 +170,17 @@ class AzureRMServers(AzureRMModuleBase):
         for key in list(self.module_arg_spec.keys()) + ['tags']:
             if hasattr(self, key):
                 setattr(self, key, kwargs[key])
-            elif key == "tags" and kwargs[key] is not None:
-                self.parameters.update({"tags": kwargs[key]})
-            elif key == "location" and kwargs[key] is not None:
-                self.parameters.update({"location": kwargs[key]})
-            elif key == "admin_username" and kwargs[key] is not None:
-                self.parameters.update({"administrator_login": kwargs[key]})
-            elif key == "admin_password" and kwargs[key] is not None:
-                self.parameters.update({"administrator_login_password": kwargs[key]})
-            elif key == "version" and kwargs[key] is not None:
-                self.parameters.update({"version": kwargs[key]})
-            elif key == "identity" and kwargs[key] is not None:
-                self.parameters.update({"identity": {"type": kwargs[key]}})
+            elif kwargs[key] is not None:
+                if key == "location":
+                    self.parameters.update({"location": kwargs[key]})
+                elif key == "admin_username":
+                    self.parameters.update({"administrator_login": kwargs[key]})
+                elif key == "admin_password":
+                    self.parameters.update({"administrator_login_password": kwargs[key]})
+                elif key == "version":
+                    self.parameters.update({"version": kwargs[key]})
+                elif key == "identity":
+                    self.parameters.update({"identity": {"type": kwargs[key]}})
 
         old_response = None
         results = dict()
@@ -230,26 +222,24 @@ class AzureRMServers(AzureRMModuleBase):
                 self.results['changed'] = True
             else:
                 self.results['changed'] = old_response.__ne__(response)
-
-            self.results.update(response)
-
-            # remove unnecessary fields from return state
-            self.results.pop('name', None)
-            self.results.pop('type', None)
-            self.results.pop('tags', None)
-            self.results.pop('location', None)
-            self.results.pop('identity', None)
-            self.results.pop('kind', None)
-            self.results.pop('administrator_login', None)
             self.log("Creation / Update done")
         elif self.to_do == Actions.Delete:
             self.log("SQL Server instance deleted")
+
+            if self.check_mode:
+                return self.results
+
             self.delete_sqlserver()
             self.results['changed'] = True
         else:
             self.log("SQL Server instance unchanged")
-            self.results['state'] = old_response
             self.results['changed'] = False
+            response = old_response
+
+        self.results["id"] = response["id"]
+        self.results["version"] = response["version"]
+        self.results["state"] = response["state"]
+        self.results["fully_qualified_domain_name"] = response["fully_qualified_domain_name"]
 
         return self.results
